requirements.txt
openai>=1.0.0
yfinance>=0.2.36
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
streamlit>=1.37.0
ddgs>=9.0.0
plotly>=5.0
